git-svn-id: http://svn.svnkit.com/repos/sqljet@744 06c6b4df-3a75-4b2a-9d74-3e57e8079485

+++ branches/1.0.x/sqljet/src/org/tmatesoft/sqljet/core/internal/fs/SqlJetFileSystem.java
@@ -224,9 +224,26 @@ public class SqlJetFileSystem implements ISqlJetFileSystem {
      */
     public boolean delete(File path, boolean sync) {
         assert (null != path);
-        final boolean delete = path.delete();
-        // TODO may be it is possible to syncing in some way
-        return delete;
+        if (path == null || !path.exists()) {
+            return false;
+        } else {
+            boolean delete = path.delete();
+            if (sync && !delete && path.exists()) {
+                for (int i = 0; i < 100; i++) {
+                    delete = path.delete();
+                    if (delete && !path.exists()) {
+                        break;
+                    } else {
+                        try {
+                            Thread.sleep(10);
+                        } catch (InterruptedException e) {
+                            break;
+                        }
+                    }
+                }
+            }
+            return delete && !path.exists();
+        }
     }
 
     /*
